[Go-Gen] Rename objects to be in line with standard
